add params sort and model comment

+++ src/services/question.service.ts
@@ -13,7 +13,19 @@ export class QuestionService {
     static async getById(id: number, userId: number) {
         return QuestionRepository.getById(id, userId)
     }
-    static async delete(id: number) {
+
+    static async checkOwnership(id: number, userId: number): Promise<boolean> {
+        const question = await this.getById(id, userId)
+        return question?.userId === userId
+    }
+    static async delete(id: number, userId: number) {
+        const isOwner = await this.checkOwnership(id, userId)
+        if(!isOwner) {
+            throw {
+                message: `Forbidden: Question: ${id} has not ${userId}`,
+                status: 403
+            }
+        }
         return QuestionRepository.deleteById(id)
     }
 }
